Initial commit. Content migration & basic config to finish style
Initial commit which brings the content present in the current Ghost
blog migrated to Hugo Markdown with frontmatter.
Initial commit configuration with some not real data to finish the
styling according the ifraixedes theme and create some specific layout and types of sections.

Added one layout and type section in this commit