refactor: Change theme
Don't use anymore my custom theme and use one of the public available
for not having to maintained and face browser related issues.

Because of changing the theme, I had to copy my favicon and a short
code that my theme has to this site. I also copied other short codes
that my theme has and can be useful to this site.

Because Twitter is now X, I renamed those short codes and updated the
content pages that used any of them.