fix: Set site params properly and remove comments
Set the site params in the right section and remove commented settings
because they were commented out when I change the theme and forgot, and
remove the Google analytics because I won't use it.

I also removed the YouTube embed short code because I realized that the
current version of Hugo has an internal short code for it.