content: Add blog post Ach Linux network config
Add a new blog post about configuring the network in a fresh Arch Linux
installation.

The post is made from my notes about installing Arch Linux on a new
laptop.

This commit also adjust the initial post about the Arch Linux base
installation to link to this new one and form a series of blog post that
I plan to publish and add blank lines on code blocks for enhance the
markdown reading and keep it consistent with the new post.